Partially refactor lunr search in cdes into its own wrapper hook on the lower-level lunr hook.
