+++ src/hooks/index.js
@@ -1,3 +1,4 @@
 export * from './use-scroll-position'
 export * from './use-local-storage'
 export * from './use-search-index'
+export * from './use-lunr-search'
